+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to
 - Add `OfferingDeepLink` model to store links of offering
   to purchase a training from an external platform
 - Handle deep link in back office
+- Create standalone to_own orders in admin
+- Expose voucher code and claim status in order admin API
+- Display voucher code in order detail view
+- Add voucher code copy-to-clipboard in orders list
 
 ### Fixed